requirements.txt
# Gestion des Invitations - Dépendances

# Interface graphique
PyQt5>=5.15.10

# Manipulation d'images et PSD
Pillow>=10.0.0
psd-tools>=1.9.0

# Génération de QR codes
qrcode[pil]>=7.4.2

# Base de données

# Génération de rapports PDF
reportlab>=4.0.0

# Communication avec Photoshop (optionnel)
# photoshop-python-api>=0.19.0

# Utilitaires
python-dateutil>=2.8.2
opencv-python>=4.8.0
pyzbar>=0.1.9
